Rebasing changes against master.
