+++ storagenode/storagenodedb/bandwidthdb.go
@@ -16,25 +16,30 @@ import (
 	"storj.io/storj/storagenode/bandwidth"
 )
 
+var ErrBandwidth = errs.Class("bandwidthdb error")
+
 type bandwidthdb struct {
 	// Moved to top of struct to resolve alignment issue with atomic operations on ARM
 	usedSpace int64
 	usedMu    sync.RWMutex
 	usedSince time.Time
 
-	*InfoDB
+	location string
+	SQLDB
 }
 
-// Bandwidth returns table for storing bandwidth usage.
-func (db *DB) Bandwidth() bandwidth.DB { return db.info.Bandwidth() }
-
-// Bandwidth returns table for storing bandwidth usage.
-func (db *InfoDB) Bandwidth() bandwidth.DB { return &db.bandwidthdb }
+// newBandwidth returns a new instance of usedSerials initialized with the specified database.
+func newBandwidth(db SQLDB, location string) *bandwidthdb {
+	return &bandwidthdb{
+		location: location,
+		SQLDB:    db,
+	}
+}
 
 // Add adds bandwidth usage to the table
 func (db *bandwidthdb) Add(ctx context.Context, satelliteID storj.NodeID, action pb.PieceAction, amount int64, created time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
-	_, err = db.db.Exec(`
+	_, err = db.Exec(`
 		INSERT INTO
 			bandwidth_usage(satellite_id, action, amount, created_at)
 		VALUES(?, ?, ?, ?)`, satelliteID, action, amount, created.UTC())
@@ -54,7 +59,7 @@ func (db *bandwidthdb) Add(ctx context.Context, satelliteID storj.NodeID, action
 			db.usedSpace = usage.Total()
 		}
 	}
-	return ErrInfo.Wrap(err)
+	return ErrBandwidth.Wrap(err)
 }
 
 // MonthSummary returns summary of the current months bandwidth usages
@@ -84,7 +89,7 @@ func (db *bandwidthdb) Summary(ctx context.Context, from, to time.Time) (_ *band
 
 	from = from.UTC()
 	to = to.UTC()
-	rows, err := db.db.Query(`
+	rows, err := db.Query(`
 		SELECT action, sum(a) amount from(
 				SELECT action, sum(amount) a
 				FROM bandwidth_usage
@@ -101,7 +106,7 @@ func (db *bandwidthdb) Summary(ctx context.Context, from, to time.Time) (_ *band
 		if err == sql.ErrNoRows {
 			return usage, nil
 		}
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrBandwidth.Wrap(err)
 	}
 	defer func() { err = errs.Combine(err, rows.Close()) }()
 
@@ -110,12 +115,12 @@ func (db *bandwidthdb) Summary(ctx context.Context, from, to time.Time) (_ *band
 		var amount int64
 		err := rows.Scan(&action, &amount)
 		if err != nil {
-			return nil, ErrInfo.Wrap(err)
+			return nil, ErrBandwidth.Wrap(err)
 		}
 		usage.Include(action, amount)
 	}
 
-	return usage, ErrInfo.Wrap(rows.Err())
+	return usage, ErrBandwidth.Wrap(rows.Err())
 }
 
 // SummaryBySatellite returns summary of bandwidth usage grouping by satellite.
@@ -126,7 +131,7 @@ func (db *bandwidthdb) SummaryBySatellite(ctx context.Context, from, to time.Tim
 
 	from = from.UTC()
 	to = to.UTC()
-	rows, err := db.db.Query(`
+	rows, err := db.Query(`
 	SELECT satellite_id, action, sum(a) amount from(
 			SELECT satellite_id, action, sum(amount) a
 			FROM bandwidth_usage
@@ -143,7 +148,7 @@ func (db *bandwidthdb) SummaryBySatellite(ctx context.Context, from, to time.Tim
 		if err == sql.ErrNoRows {
 			return entries, nil
 		}
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrBandwidth.Wrap(err)
 	}
 	defer func() { err = errs.Combine(err, rows.Close()) }()
 
@@ -154,7 +159,7 @@ func (db *bandwidthdb) SummaryBySatellite(ctx context.Context, from, to time.Tim
 
 		err := rows.Scan(&satelliteID, &action, &amount)
 		if err != nil {
-			return nil, ErrInfo.Wrap(err)
+			return nil, ErrBandwidth.Wrap(err)
 		}
 
 		entry, ok := entries[satelliteID]
@@ -166,7 +171,7 @@ func (db *bandwidthdb) SummaryBySatellite(ctx context.Context, from, to time.Tim
 		entry.Include(action, amount)
 	}
 
-	return entries, ErrInfo.Wrap(rows.Err())
+	return entries, ErrBandwidth.Wrap(rows.Err())
 }
 
 // Rollup bandwidth_usage data earlier than the current hour, then delete the rolled up records
@@ -180,7 +185,7 @@ func (db *bandwidthdb) Rollup(ctx context.Context) (err error) {
 
 	tx, err := db.Begin()
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrBandwidth.Wrap(err)
 	}
 
 	defer func() {
@@ -203,12 +208,12 @@ func (db *bandwidthdb) Rollup(ctx context.Context) (err error) {
 		DELETE FROM bandwidth_usage WHERE datetime(created_at) < datetime(?);
 	`, hour, hour)
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrBandwidth.Wrap(err)
 	}
 
 	_, err = result.RowsAffected()
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrBandwidth.Wrap(err)
 	}
 
 	return nil
