+++ storagenode/storagenodedb/pieceexpiration.go
@@ -13,21 +13,26 @@ import (
 	"storj.io/storj/storagenode/pieces"
 )
 
+var ErrPieceExpiration = errs.Class("piece expiration error")
+
 type pieceExpirationDB struct {
-	*InfoDB
+	location string
+	SQLDB
 }
 
-// PieceExpirationDB returns database for storing piece expiration data
-func (db *DB) PieceExpirationDB() pieces.PieceExpirationDB { return db.info.PieceExpirationDB() }
-
-// PieceExpirationDB returns database for storing piece expiration data
-func (db *InfoDB) PieceExpirationDB() pieces.PieceExpirationDB { return &db.pieceExpirationDB }
+// newPieceExpirationDB returns a new instance of pieceExpirationDB initialized with the specified database.
+func newPieceExpirationDB(db SQLDB, location string) *pieceExpirationDB {
+	return &pieceExpirationDB{
+		location: location,
+		SQLDB:    db,
+	}
+}
 
 // GetExpired gets piece IDs that expire or have expired before the given time
 func (db *pieceExpirationDB) GetExpired(ctx context.Context, expiresBefore time.Time, limit int64) (expiredPieceIDs []pieces.ExpiredInfo, err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	rows, err := db.db.QueryContext(ctx, `
+	rows, err := db.QueryContext(ctx, `
 		SELECT satellite_id, piece_id
 			FROM piece_expirations
 			WHERE piece_expiration < ?
@@ -35,7 +40,7 @@ func (db *pieceExpirationDB) GetExpired(ctx context.Context, expiresBefore time.
 			LIMIT ?
 	`, expiresBefore.UTC(), expiresBefore.UTC(), limit)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceExpiration.Wrap(err)
 	}
 	defer func() { err = errs.Combine(err, rows.Close()) }()
 
@@ -44,7 +49,7 @@ func (db *pieceExpirationDB) GetExpired(ctx context.Context, expiresBefore time.
 		var pieceID storj.PieceID
 		err = rows.Scan(&satelliteID, &pieceID)
 		if err != nil {
-			return nil, ErrInfo.Wrap(err)
+			return nil, ErrPieceExpiration.Wrap(err)
 		}
 		expiredPieceIDs = append(expiredPieceIDs, pieces.ExpiredInfo{
 			SatelliteID: satelliteID,
@@ -59,18 +64,18 @@ func (db *pieceExpirationDB) GetExpired(ctx context.Context, expiresBefore time.
 func (db *pieceExpirationDB) SetExpiration(ctx context.Context, satellite storj.NodeID, pieceID storj.PieceID, expiresAt time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.ExecContext(ctx, `
+	_, err = db.ExecContext(ctx, `
 		INSERT INTO piece_expirations(satellite_id, piece_id, piece_expiration)
 			VALUES (?,?,?)
 	`, satellite, pieceID, expiresAt.UTC())
-	return ErrInfo.Wrap(err)
+	return ErrPieceExpiration.Wrap(err)
 }
 
 // DeleteExpiration removes an expiration record for the given piece ID on the given satellite
 func (db *pieceExpirationDB) DeleteExpiration(ctx context.Context, satelliteID storj.NodeID, pieceID storj.PieceID) (found bool, err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	result, err := db.db.ExecContext(ctx, `
+	result, err := db.ExecContext(ctx, `
 		DELETE FROM piece_expirations
 			WHERE satellite_id = ? AND piece_id = ?
 	`, satelliteID, pieceID)
@@ -89,11 +94,11 @@ func (db *pieceExpirationDB) DeleteExpiration(ctx context.Context, satelliteID s
 func (db *pieceExpirationDB) DeleteFailed(ctx context.Context, satelliteID storj.NodeID, pieceID storj.PieceID, when time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.ExecContext(ctx, `
+	_, err = db.ExecContext(ctx, `
 		UPDATE piece_expirations
 			SET deletion_failed_at = ?
 			WHERE satellite_id = ?
 				AND piece_id = ?
 	`, when.UTC(), satelliteID, pieceID)
-	return ErrInfo.Wrap(err)
+	return ErrPieceExpiration.Wrap(err)
 }
