+++ storagenode/storagenodedb/pieceinfo.go
@@ -18,15 +18,20 @@ import (
 	"storj.io/storj/storagenode/pieces"
 )
 
+var ErrPieceInfo = errs.Class("v0pieceinfo error")
+
 type v0PieceInfo struct {
-	*InfoDB
+	location string
+	SQLDB
 }
 
-// V0PieceInfo returns database for storing piece information
-func (db *DB) V0PieceInfo() pieces.V0PieceInfoDB { return db.info.V0PieceInfo() }
-
-// V0PieceInfo returns database for storing piece information
-func (db *InfoDB) V0PieceInfo() pieces.V0PieceInfoDB { return &db.v0PieceInfo }
+// newPieceInfo returns a new instance of pieceinfo initialized with the specified database.
+func newPieceInfo(db SQLDB, location string) *v0PieceInfo {
+	return &v0PieceInfo{
+		location: location,
+		SQLDB:    db,
+	}
+}
 
 // Add inserts piece information into the database.
 func (db *v0PieceInfo) Add(ctx context.Context, info *pieces.Info) (err error) {
@@ -34,12 +39,12 @@ func (db *v0PieceInfo) Add(ctx context.Context, info *pieces.Info) (err error) {
 
 	orderLimit, err := proto.Marshal(info.OrderLimit)
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrPieceInfo.Wrap(err)
 	}
 
 	uplinkPieceHash, err := proto.Marshal(info.UplinkPieceHash)
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrPieceInfo.Wrap(err)
 	}
 
 	var pieceExpiration *time.Time
@@ -49,24 +54,24 @@ func (db *v0PieceInfo) Add(ctx context.Context, info *pieces.Info) (err error) {
 	}
 
 	// TODO remove `uplink_cert_id` from DB
-	_, err = db.db.ExecContext(ctx, db.Rebind(`
+	_, err = db.ExecContext(ctx, `
 		INSERT INTO
 			pieceinfo_(satellite_id, piece_id, piece_size, piece_creation, piece_expiration, order_limit, uplink_piece_hash, uplink_cert_id)
 		VALUES (?,?,?,?,?,?,?,?)
-	`), info.SatelliteID, info.PieceID, info.PieceSize, info.PieceCreation.UTC(), pieceExpiration, orderLimit, uplinkPieceHash, 0)
+	`, info.SatelliteID, info.PieceID, info.PieceSize, info.PieceCreation.UTC(), pieceExpiration, orderLimit, uplinkPieceHash, 0)
 
-	return ErrInfo.Wrap(err)
+	return ErrPieceInfo.Wrap(err)
 }
 
 func (db *v0PieceInfo) getAllPiecesOwnedBy(ctx context.Context, blobStore storage.Blobs, satelliteID storj.NodeID) ([]v0StoredPieceAccess, error) {
-	rows, err := db.db.QueryContext(ctx, db.Rebind(`
+	rows, err := db.QueryContext(ctx, `
 		SELECT piece_id, piece_size, piece_creation, piece_expiration
 		FROM pieceinfo_
 		WHERE satellite_id = ?
 		ORDER BY piece_id
-	`), satelliteID)
+	`, satelliteID)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceInfo.Wrap(err)
 	}
 	defer func() { err = errs.Combine(err, rows.Close()) }()
 	var pieceInfos []v0StoredPieceAccess
@@ -78,7 +83,7 @@ func (db *v0PieceInfo) getAllPiecesOwnedBy(ctx context.Context, blobStore storag
 		thisAccess := &pieceInfos[len(pieceInfos)-1]
 		err = rows.Scan(&thisAccess.pieceID, &thisAccess.pieceSize, &thisAccess.creationTime, &thisAccess.expirationTime)
 		if err != nil {
-			return nil, ErrInfo.Wrap(err)
+			return nil, ErrPieceInfo.Wrap(err)
 		}
 	}
 	return pieceInfos, nil
@@ -123,13 +128,13 @@ func (db *v0PieceInfo) Get(ctx context.Context, satelliteID storj.NodeID, pieceI
 	var uplinkPieceHash []byte
 	var pieceExpiration *time.Time
 
-	err = db.db.QueryRowContext(ctx, db.Rebind(`
+	err = db.QueryRowContext(ctx, `
 		SELECT piece_size, piece_creation, piece_expiration, order_limit, uplink_piece_hash
 		FROM pieceinfo_
 		WHERE satellite_id = ? AND piece_id = ?
-	`), satelliteID, pieceID).Scan(&info.PieceSize, &info.PieceCreation, &pieceExpiration, &orderLimit, &uplinkPieceHash)
+	`, satelliteID, pieceID).Scan(&info.PieceSize, &info.PieceCreation, &pieceExpiration, &orderLimit, &uplinkPieceHash)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceInfo.Wrap(err)
 	}
 
 	if pieceExpiration != nil {
@@ -139,13 +144,13 @@ func (db *v0PieceInfo) Get(ctx context.Context, satelliteID storj.NodeID, pieceI
 	info.OrderLimit = &pb.OrderLimit{}
 	err = proto.Unmarshal(orderLimit, info.OrderLimit)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceInfo.Wrap(err)
 	}
 
 	info.UplinkPieceHash = &pb.PieceHash{}
 	err = proto.Unmarshal(uplinkPieceHash, info.UplinkPieceHash)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceInfo.Wrap(err)
 	}
 
 	return info, nil
@@ -155,34 +160,34 @@ func (db *v0PieceInfo) Get(ctx context.Context, satelliteID storj.NodeID, pieceI
 func (db *v0PieceInfo) Delete(ctx context.Context, satelliteID storj.NodeID, pieceID storj.PieceID) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.ExecContext(ctx, db.Rebind(`
+	_, err = db.ExecContext(ctx, `
 		DELETE FROM pieceinfo_
 		WHERE satellite_id = ?
 		  AND piece_id = ?
-	`), satelliteID, pieceID)
+	`, satelliteID, pieceID)
 
-	return ErrInfo.Wrap(err)
+	return ErrPieceInfo.Wrap(err)
 }
 
 // DeleteFailed marks piece as a failed deletion.
 func (db *v0PieceInfo) DeleteFailed(ctx context.Context, satelliteID storj.NodeID, pieceID storj.PieceID, now time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.ExecContext(ctx, db.Rebind(`
+	_, err = db.ExecContext(ctx, `
 		UPDATE pieceinfo_
 		SET deletion_failed_at = ?
 		WHERE satellite_id = ?
 		  AND piece_id = ?
-	`), now.UTC(), satelliteID, pieceID)
+	`, now.UTC(), satelliteID, pieceID)
 
-	return ErrInfo.Wrap(err)
+	return ErrPieceInfo.Wrap(err)
 }
 
 // GetExpired gets ExpiredInfo records for pieces that are expired.
 func (db *v0PieceInfo) GetExpired(ctx context.Context, expiredAt time.Time, limit int64) (infos []pieces.ExpiredInfo, err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	rows, err := db.db.QueryContext(ctx, db.Rebind(`
+	rows, err := db.QueryContext(ctx, `
 		SELECT satellite_id, piece_id
 		FROM pieceinfo_
 		WHERE piece_expiration IS NOT NULL
@@ -190,16 +195,16 @@ func (db *v0PieceInfo) GetExpired(ctx context.Context, expiredAt time.Time, limi
 		AND ((deletion_failed_at IS NULL) OR deletion_failed_at <> ?)
 		ORDER BY satellite_id
 		LIMIT ?
-	`), expiredAt.UTC(), expiredAt.UTC(), limit)
+	`, expiredAt.UTC(), expiredAt.UTC(), limit)
 	if err != nil {
-		return nil, ErrInfo.Wrap(err)
+		return nil, ErrPieceInfo.Wrap(err)
 	}
 	defer func() { err = errs.Combine(err, rows.Close()) }()
 	for rows.Next() {
 		info := pieces.ExpiredInfo{InPieceInfo: true}
 		err = rows.Scan(&info.SatelliteID, &info.PieceID)
 		if err != nil {
-			return infos, ErrInfo.Wrap(err)
+			return infos, ErrPieceInfo.Wrap(err)
 		}
 		infos = append(infos, info)
 	}
