+++ storagenode/storagenodedb/usedserials.go
@@ -13,34 +13,39 @@ import (
 	"storj.io/storj/storagenode/piecestore"
 )
 
+var ErrUsedSerials = errs.Class("usedserials error")
+
 type usedSerials struct {
-	*InfoDB
+	location string
+	SQLDB
 }
 
-// UsedSerials returns used serials database.
-func (db *DB) UsedSerials() piecestore.UsedSerials { return db.info.UsedSerials() }
-
-// UsedSerials returns used serials database.
-func (db *InfoDB) UsedSerials() piecestore.UsedSerials { return &usedSerials{db} }
+// newUsedSerials returns a new instance of usedSerials initialized with the specified database.
+func newUsedSerials(db SQLDB, location string) *usedSerials {
+	return &usedSerials{
+		location: location,
+		SQLDB:    db,
+	}
+}
 
 // Add adds a serial to the database.
 func (db *usedSerials) Add(ctx context.Context, satelliteID storj.NodeID, serialNumber storj.SerialNumber, expiration time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.Exec(`
+	_, err = db.Exec(`
 		INSERT INTO
 			used_serial_(satellite_id, serial_number, expiration)
 		VALUES(?, ?, ?)`, satelliteID, serialNumber, expiration.UTC())
 
-	return ErrInfo.Wrap(err)
+	return ErrUsedSerials.Wrap(err)
 }
 
 // DeleteExpired deletes expired serial numbers
 func (db *usedSerials) DeleteExpired(ctx context.Context, now time.Time) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	_, err = db.db.Exec(`DELETE FROM used_serial_ WHERE expiration < ?`, now.UTC())
-	return ErrInfo.Wrap(err)
+	_, err = db.Exec(`DELETE FROM used_serial_ WHERE expiration < ?`, now.UTC())
+	return ErrUsedSerials.Wrap(err)
 }
 
 // IterateAll iterates all serials.
@@ -48,11 +53,11 @@ func (db *usedSerials) DeleteExpired(ctx context.Context, now time.Time) (err er
 func (db *usedSerials) IterateAll(ctx context.Context, fn piecestore.SerialNumberFn) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	rows, err := db.db.Query(`SELECT satellite_id, serial_number, expiration FROM used_serial_`)
+	rows, err := db.Query(`SELECT satellite_id, serial_number, expiration FROM used_serial_`)
 	if err != nil {
-		return ErrInfo.Wrap(err)
+		return ErrUsedSerials.Wrap(err)
 	}
-	defer func() { err = errs.Combine(err, ErrInfo.Wrap(rows.Close())) }()
+	defer func() { err = errs.Combine(err, ErrUsedSerials.Wrap(rows.Close())) }()
 
 	for rows.Next() {
 		var satelliteID storj.NodeID
@@ -61,11 +66,11 @@ func (db *usedSerials) IterateAll(ctx context.Context, fn piecestore.SerialNumbe
 
 		err := rows.Scan(&satelliteID, &serialNumber, &expiration)
 		if err != nil {
-			return ErrInfo.Wrap(err)
+			return ErrUsedSerials.Wrap(err)
 		}
 
 		fn(satelliteID, serialNumber, expiration)
 	}
 
-	return ErrInfo.Wrap(rows.Err())
+	return ErrUsedSerials.Wrap(rows.Err())
 }
